Support parallelizing the packument mirror
A naive approach to parallelization would cause problems with the
seq number incrementing. You run the risk of completing a higher
number sequence id before a lower number. You could guard against
the seq regressing by only updating the seq value if the new value
is greater - but this actually has the opposite effect we want! The
real risk is that we update the seq value to a number higher than
some currently inflight requests, and then the process crashes.
When we resume downloading - we won't retry these partial requests,
we will move on and they'll be missed.

To address this we used cargo to batch requests, and we only
increment seq after the entire batch completes. This might cause
a slight amount of rework when the process restarts but it strongly
guards against missing any packages.